[BACKPORT][wayland gpu] Use only announced buffer formats.
It turned out that it was not a good idea to try to guess buffer
formats if a Wayland compositor had not announced supported buffer
formats.

If such a case happened (the Wayland compositor did not announce the
formats), guessing them resulted in crashes in YouTube videos
for example.

Thus, it's better to fallback to software GpuMemoryBuffers rather
than crashing.

What is more, the Wayland spec says it must announce the supported
formats once zwp_linux_dmabuf interface is bound, but in some versions
of the compositors it does not happen. It looks more like a bug, which
must be reported.

Bug: 890284
Change-Id: I9bd20019709b0cbd82fa9654627b825adceb479b
Reviewed-on: https://chromium-review.googlesource.com/1253641
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#595436}