[BACKPORT] [ozone/wayland] Fix crash when there is no zwp_linux_dmabuf
This fixes the case when zwp_linux_dmabuf interface is not available.

The fix is to make WaylandConnectionProxy to reset gbm_device, and
return nullptr instead. This makes the WaylandSurfaceFactory return
nullptr for SurfacelessViewGLSurface, which results in a software
rendering path usage when the "--in-process-gpu" flag is not
passed and ViewGLSurface, when the flag is passed.

TL;DR:

This change results in the following changes in two different cases:
  1) in-process-gpu is passed: in that mode, WaylandSurfaceFactory
tries to create surfaceless gl surface, which uses gbm and linux dmabuf
interface to create buffers. If neither gbm nor the interface are available,
surfaceless gl surface can't be created and nullptr is returned. Then,
Chromium asks to create ViewGlSurface, which uses wl_egl_surface.
  2) the flag is not passed: in that mode, the factory creates the
above mentioned surfaceless gl surface, which uses gbm and linux dmabuf
interface. If those are not available, nullptr is returned and chromium
falls back to software rendering (please note that it's impossible to create
wl_egl_surface, because it requires accessing wl_surface on the browser
process side). Note, software rendering CL is under review at
crrev.com/c/1454623

Bug: 932098
Change-Id: I95f88018ded48e7a18864f957137c2b478e788ac
Reviewed-on: https://chromium-review.googlesource.com/c/1477591
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#635044}