[open][upstreamable][ozone] Report hotplug devices
:Release Notes:
Report to data device manager the hotplug devices.

:Detailed Notes:
Some code in chromium depends on detection of the input devices using
DataDeviceManager. But ozone wayland backend was not populating that
information. This patch implements basic reporting of available devices,
as wayland is not giving detailed information.

:Testing Performed:
maxTouchPoints is set to 1.
Passing on command line --ignore-max-touch-points sets maxTouchPoints to 0.
Passing on command line --force-max-touch-points=10 sets maxTouchPoints to 10.

:QA Notes:
Test-cases: QS-8196
Depends on: http://gpro.lge.com/259057

:Issues Addressed:
[PLAT-86381] [Signage]: List of items does not scrolls *Up with each swipe.*

Change-Id: Idb707f0be3a18a793001e9476f95c0b11d6b4d04
Reviewed-on: http://gpro.lge.com/258488
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>